tests: toyserver: create ToyNetwork for tests with ToyServer
Create a shared `ToyNetwork` which can be used for unittests depending
on the toyserver.
